requirements.txt
requests
ics
python-dateutil